refactor: 역량 키워드 Repository 인터페이스 위치 변경 누락 수정(#52)
