--- Domain-Module/src/main/kotlin/com/bamyanggang/domainmodule/domain/experience/repository/StrongPointRepository.kt
@@ -1,9 +1,0 @@
-package com.bamyanggang.domainmodule.domain.experience.repository
-
-import com.bamyanggang.domainmodule.domain.strongpoint.aggregate.StrongPoint
-import java.util.*
-
-interface StrongPointRepository {
-    fun save(newStrongPoint: StrongPoint): UUID
-    fun findAllByUserId(userId: UUID): List<StrongPoint>
-}
