+++ Domain-Module/src/main/kotlin/com/bamyanggang/domainmodule/domain/strongpoint/repository/StrongPointRepository.kt
@@ -1,4 +1,9 @@
 package com.bamyanggang.domainmodule.domain.strongpoint.repository
 
+import com.bamyanggang.domainmodule.domain.strongpoint.aggregate.StrongPoint
+import java.util.*
+
 interface StrongPointRepository {
+    fun save(newStrongPoint: StrongPoint): UUID
+    fun findAllByUserId(userId: UUID): List<StrongPoint>
 }
